refactor(tools): rename legacy_migrate to migrate_v2
Per PR review: legacy_migrate suggested erigon→op data migration; the tool
actually migrates the v1 to v2 storage layout.
